requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
pyodbc
azure-storage-blob
python-dotenv